# Frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/enrollments/{activityId}/{memberId}": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Enrollment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enrollment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enrollment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Enrollments"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Enrollments"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/payments/enrollment/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPayment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPayment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPayment"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/studyenrollments": {
      "get": {
        "tags": [
          "StudyEnrollments"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyEnrollment"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyEnrollment"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyEnrollment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StudyEnrollments"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PostStudyEnrollmentDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostStudyEnrollmentDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostStudyEnrollmentDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostStudyEnrollmentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              }
            }
          }
        }
      }
    },
    "/api/studyenrollments/{id}": {
      "get": {
        "tags": [
          "StudyEnrollments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyEnrollment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StudyEnrollments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Activity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "posterFileName": {
            "type": "string",
            "nullable": true
          },
          "posterPath": {
            "type": "string",
            "nullable": true
          },
          "dutchDescription": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "englishDescription": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "dateTimeStart": {
            "type": "string",
            "format": "date-time"
          },
          "dateTimeEnd": {
            "type": "string",
            "format": "date-time"
          },
          "unenrollmentDeadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "enrollmentDeadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "location": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "participantLimit": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organizerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organizer": {
            "$ref": "#/components/schemas/Group"
          },
          "specificationQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecificationQuestion"
            },
            "nullable": true
          },
          "showInKoala": {
            "type": "boolean"
          },
          "showOnWebsite": {
            "type": "boolean"
          },
          "isEnrollable": {
            "type": "boolean"
          },
          "areParticipantsVisible": {
            "type": "boolean"
          },
          "isAdultOnly": {
            "type": "boolean"
          },
          "allowedAudience": {
            "$ref": "#/components/schemas/TargetAudience"
          },
          "isOpenForPayment": {
            "type": "boolean"
          },
          "vatRate": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "nullable": true
          },
          "glAccountId": {
            "type": "string",
            "nullable": true
          },
          "costCenterId": {
            "type": "string",
            "nullable": true
          },
          "costUnitId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "content": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Member"
          }
        },
        "additionalProperties": false
      },
      "Enrollment": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "EnrollmentPayment": {
        "required": [
          "mollieId",
          "paymentIntentUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "mollieId": {
            "type": "string",
            "nullable": true
          },
          "paymentIntentUrl": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "type": "string",
            "nullable": true
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          }
        },
        "additionalProperties": false
      },
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "groupMemberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMembership"
            },
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/GroupType"
          }
        },
        "additionalProperties": false
      },
      "GroupMembership": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "group": {
            "$ref": "#/components/schemas/Group"
          },
          "membershipYear": {
            "type": "integer",
            "format": "int32"
          },
          "roleAliasId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "roleAlias": {
            "$ref": "#/components/schemas/RoleAlias"
          }
        },
        "additionalProperties": false
      },
...
      "GroupType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Language": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "Member": {
        "required": [
          "address",
          "email",
          "firstName",
          "lastName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "keycloakId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "studentNumber": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "address": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "registeredOn": {
            "type": "string",
            "format": "date-time"
          },
          "studyEnrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudyEnrollment"
            },
            "nullable": true
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "nullable": true
          },
          "groupMemberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMembership"
            },
            "nullable": true
          },
          "preferredLanguage": {
            "$ref": "#/components/schemas/Language"
          },
          "gratie": {
            "type": "boolean"
          },
          "lidVanVerdienste": {
            "type": "boolean"
          },
          "ereLid": {
            "type": "boolean"
          },
          "begunstiger": {
            "type": "boolean"
          },
          "suspended": {
            "type": "boolean"
          },
          "announcements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Announcement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Operation": {
        "type": "object",
        "properties": {
          "operationType": {
            "$ref": "#/components/schemas/OperationType"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OperationType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PostStudyEnrollmentDTO": {
        "required": [
          "enrollmentDate",
          "memberId",
          "studyId"
        ],
        "type": "object",
        "properties": {
          "studyId": {
            "type": "integer",
            "format": "int32"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "enrollmentDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/StudyStatus"
          }
        },
        "additionalProperties": false
      },
      "QuestionType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoleAlias": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SpecificationQuestion": {
        "required": [
          "activity",
          "questionDutch",
          "questionEnglish"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "questionDutch": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "questionEnglish": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "isMandatory": {
            "type": "boolean"
          },
          "isPublic": {
            "type": "boolean"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          }
        },
        "additionalProperties": false
      },
...
      "Study": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nominalDurationYears": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/StudyType"
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudyEnrollment"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StudyEnrollment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "studyId": {
            "type": "integer",
            "format": "int32"
          },
          "study": {
            "$ref": "#/components/schemas/Study"
          },
          "enrollmentDate": {
            "type": "string",
            "format": "date-time"
          },
          "completionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/StudyStatus"
          }
        },
        "additionalProperties": false
      },
...
      "StudyStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "StudyType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "TargetAudience": {
        "enum": [
          0,
          1,
          2,
          4,
          8,
          15
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste your JWT token here",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}